+++ R/S4.R
@@ -33,11 +33,47 @@ S4_to_R7_class <- function(x, error_base = "") {
   }
 }
 
-# R7 handles unions at method registration time, where as S4 handles them at
-# dispatch time.
-S4_strip_union <- function(class_names) {
-  classes <- lapply(class_names, methods::getClass)
-  is_union <- vlapply(classes, methods::is, "ClassUnionRepresentation")
+S4_class_dispatch <- function(x) {
+  x <- methods::getClass(x)
+  self <- S4_class_name(x)
 
-  setdiff(class_names[!is_union], "vector")
+  # Find class objects for super classes
+  extends <- unname(methods::extends(x, fullInfo = TRUE))
+  extends <- Filter(function(x) methods::is(x, "SClassExtension"), extends)
+  classes <- lapply(extends, function(x) methods::getClass(x@superClass))
+
+  # Remove virtual classes that aren't S3. This removes unions because R7
+  # handles them in method registration, not dispatch.
+  classes <- Filter(function(x) !x@virtual || is_oldClass(x), classes)
+
+  c(self, vcapply(classes, S4_class_name))
 }
+
+is_oldClass <- function(x) {
+  x@virtual && methods::extends(x, "oldClass") && x@className != "oldClass"
+}
+
+S4_class_name <- function(x) {
+  if (is_oldClass(x)) {
+    return(x@className)
+  }
+
+  class <- x@className
+  package <- x@package %||% attr(class, "package")
+
+  if (identical(package, "methods") && class %in% names(base_classes)) {
+    class
+  } else if (is.null(package) || identical(package, ".GlobalEnv")) {
+    paste0("S4/", class)
+  } else {
+    paste0("S4/", package, "::", class)
+  }
+}
+
+S4_remove_classes <- function(classes, where = globalenv()) {
+  for (class in classes) {
+    methods::removeClass(class, topenv(where))
+  }
+}
+
+globalVariables(c("superClass", "virtual"))